LayoutTests/http/tests/websocket/tests/hybi/client-close-expected.txt

WebSocket: Test client-initiated close.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
Received: close_frame[:2]='\x88\x80'
Closed
PASS receivedMessage is "close_frame[:2]='\\x88\\x80'"
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
